refactor(server): move bridge routes into package
